+++ LojasNews/src/dao/IGerenciamentoDAO.java
@@ -5,8 +5,8 @@ public interface IGerenciamentoDAO {
     
     public boolean inserir();
     
-    public boolean atualizar (int codigo);
+    public boolean atualizar (Long codigo);
     
-    public boolean excluir (int codigo);
+    public boolean excluir (Long codigo);
     
 }
